+++ src/main/java/dandi/dandi/weatherbatch/application/port/in/WeatherBatchRequest.java
@@ -1,4 +1,4 @@
-package dandi.dandi.batch.weather.application.port.in;
+package dandi.dandi.weatherbatch.application.port.in;
 
 public class WeatherBatchRequest {
