+++ src/main/java/dandi/dandi/weatherbatch/application/service/WeatherBatchScheduler.java
@@ -1,6 +1,6 @@
-package dandi.dandi.batch.weather.application.service;
+package dandi.dandi.weatherbatch.application.service;
 
-import dandi.dandi.batch.weather.application.port.in.WeatherBatchRequest;
+import dandi.dandi.weatherbatch.application.port.in.WeatherBatchRequest;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
